+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Updated citation information for the package (#89)
 * Updated package logo to reflect the recent inclusion of patterns (#90)
+* The `pos` argument for `coord_geo()` is now recycled when `dat` is a list, removing the need for repeated `pos` values when adding multiple time scales to the same side (#86)
 
 # deeptime 2.2.0